Fix viewport on Inform7 Standard template.
In Inform7-6M62, the Standard template is incompatible with two of the
styles introduced in 71fc32c4bc0bd8d264c57c9196f611b9221d3756:

* "width: 100%" because the interpreter runs in div#gameport,
  whose width is the window width minus 15.4em;
* "overflow: hidden" because, at least the way ifvms.js does it now,
  nothing inside div#gameport has overflow: scroll or auto so this
  would break scrolling, which I think is the symptom seen in #37.

So override those styles in an embedded stylesheet in the HTML head
material provided to Inform 7.

Longer-term, if Inform 7 wants to use quixe through parchment rather
than as a separate template, it'll need to rename #gameport because the
id collides with an internal div used by quixe. (That or parchment can
rename gameport in quixe-runner.js, in (a) the HTML fragment and (b)
adding gameport: "whatever_the_new_name_is" to the set_page_title line;
but making that change would surprise anyone who has customized a
template with styles on #gameport.)